Move App into separate file
Also fixes a bug where surface wasn't configured before presentation
