+++ src/frontend/index.html
@@ -20,6 +20,7 @@
       #map_canvas> canvas {
         width:1000px;
         height:1000px;
+        contain: size;
       }
     </style>
   </head>
